Fix build and docstring warnings in OS X
After switching my build environment to use 64 bit binaries under OS X,
some code fails to build due to integer size type differences.  Attempt
to fix this by applying a little hack.

Also correct a docstring problem just introduced that was not caught
by NetBSD's Doxygen old version...

This change was r308 in Subversion.

+++ store/transaction_test.cpp
@@ -600,8 +600,9 @@ ATF_TEST_CASE_BODY(put_test_program__plain)
         ATF_REQUIRE(stmt.step());
         ATF_REQUIRE_EQ(test_program_id,
                        stmt.safe_column_int64("test_program_id"));
-        ATF_REQUIRE_EQ(test_program.timeout().to_useconds(),
-                       stmt.safe_column_int64("timeout"));
+        ATF_REQUIRE_EQ(
+            static_cast< int64_t >(test_program.timeout().to_useconds()),
+            stmt.safe_column_int64("timeout"));
     }
 }
